# backend/docs/swagger.yaml
basePath: /
definitions:
  datauptwo_app_dto.AdminEnableMFAReq:
    properties:
      code:
        type: string
      userId:
        type: integer
    required:
    - code
    - userId
    type: object
...
  datauptwo_app_dto.ResetMFAReq:
    properties:
      userId:
        type: integer
    required:
    - userId
    type: object
  datauptwo_app_dto.ResetPasswordReq:
    properties:
      newPassword:
        maxLength: 128
        minLength: 6
        type: string
      userId:
        type: integer
    required:
    - newPassword
    - userId
    type: object
  datauptwo_app_dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  datauptwo_app_dto.UserCreateReq:
    properties:
      email:
        type: string
      groupId:
        type: integer
      mfaEnabled:
        type: boolean
      nickname:
        maxLength: 64
        type: string
      password:
        maxLength: 128
        minLength: 6
        type: string
      phone:
        maxLength: 32
        type: string
      roleId:
        type: integer
      status:
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - password
    - username
    type: object
...
  datauptwo_app_dto.UserUpdateReq:
    properties:
      email:
        type: string
      groupId:
        type: integer
      id:
        type: integer
      mfaEnabled:
        type: boolean
      nickname:
        maxLength: 64
        type: string
      phone:
        maxLength: 32
        type: string
      roleId:
        type: integer
      status:
        type: string
    required:
    - id
    type: object
...
host: localhost:18421
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: datauptwo 数据管理平台 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: datauptwo API
  version: "1.0"
paths:
...
  /api/users:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 搜索关键词
        in: query
        name: keyword
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 角色ID
        in: query
        name: roleId
        type: integer
      - description: 用户组ID
        in: query
        name: groupId
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 用户列表
      tags:
      - User
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/datauptwo_app_dto.UserCreateReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 创建用户
      tags:
      - User
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/datauptwo_app_dto.UserUpdateReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 更新用户
      tags:
      - User
  /api/users/{id}:
    delete:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 删除用户
      tags:
      - User
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 获取用户详情
      tags:
      - User
  /api/users/admin-enable-mfa:
    post:
      consumes:
      - application/json
      parameters:
      - description: 启用MFA信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/datauptwo_app_dto.AdminEnableMFAReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 管理员为用户启用MFA
      tags:
      - User
  /api/users/all:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 获取所有用户
      tags:
      - User
  /api/users/generate-mfa-secret:
    get:
      parameters:
      - description: 用户ID
        in: query
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 生成MFA密钥
      tags:
      - User
  /api/users/reset-mfa:
    post:
      consumes:
      - application/json
      parameters:
      - description: 重置MFA信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/datauptwo_app_dto.ResetMFAReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 重置用户MFA
      tags:
      - User
  /api/users/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: 重置密码信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/datauptwo_app_dto.ResetPasswordReq'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/datauptwo_app_dto.Response'
      security:
      - Bearer: []
      summary: 重置用户密码
      tags:
      - User
...
swagger: "2.0"